refactor : refactor to use Traits
